requirements.txt
appdirs==1.4.3
dj-database-url==0.4.2
Django==1.11.1
django-bootstrap3==8.2.3
django-braces==1.8.1
django-crispy-forms==1.6.1
gunicorn==19.7.1
olefile==0.44
packaging==16.8
Pillow==4.1.1
psycopg2==2.7.1
pyparsing==2.2.0
pytz==2017.2
selenium==3.4.1
six==1.10.0
